Move vo_bit module under the metadata module
