+++ src/policy/marksweepspace/native_ms/global.rs
@@ -108,12 +108,12 @@ impl<VM: VMBinding> SFT for MarkSweepSpace<VM> {
 
     fn initialize_object_metadata(&self, _object: crate::util::ObjectReference, _alloc: bool) {
         #[cfg(feature = "vo_bit")]
-        crate::util::vo_bit::set_vo_bit::<VM>(_object);
+        crate::util::metadata::vo_bit::set_vo_bit::<VM>(_object);
     }
 
     #[cfg(feature = "is_mmtk_object")]
     fn is_mmtk_object(&self, addr: Address) -> bool {
-        crate::util::vo_bit::is_vo_bit_set_for_addr::<VM>(addr).is_some()
+        crate::util::metadata::vo_bit::is_vo_bit_set_for_addr::<VM>(addr).is_some()
     }
 
     fn sft_trace_object(
@@ -284,7 +284,7 @@ impl<VM: VMBinding> MarkSweepSpace<VM> {
             metadata_spec.set_zero_atomic(block.start(), Ordering::SeqCst);
         }
         #[cfg(feature = "vo_bit")]
-        crate::util::vo_bit::bzero_vo_bit(block.start(), Block::BYTES);
+        crate::util::metadata::vo_bit::bzero_vo_bit(block.start(), Block::BYTES);
     }
 
     pub fn acquire_block(&self, tls: VMThread, size: usize, align: usize) -> BlockAcquireResult {
