+++ src/util/metadata/mod.rs
@@ -227,5 +227,6 @@ pub use metadata_val_traits::*;
 pub(crate) mod log_bit;
 pub(crate) mod mark_bit;
 pub(crate) mod pin_bit;
+pub(crate) mod vo_bit;
 
 pub use global::*;
