+++ src/util/mod.rs
@@ -29,8 +29,6 @@ pub mod options;
 pub mod reference_processor;
 
 // The following modules are only public in the mmtk crate. They should only be used in MMTk core.
-/// Valid object bit
-pub(crate) mod vo_bit;
 /// An analysis framework for collecting data and profiling in GC.
 #[cfg(feature = "analysis")]
 pub(crate) mod analysis;
